Refactor hash files: hashBitstream() extracted into utils/hashBase.c to always process complete (64-bit) words at a time. Some constants' hashes updated. Also create common function hashValue64() in utils/hash64.c which each other function calls inside loops

+++ tests/src/encode/zfpEncodeBlockBase.c
@@ -107,6 +107,6 @@ _catFunc3(given_, DIM_INT_STR, Block_when_EncodeBlock_expect_BitstreamChecksumMa
   _t2(zfp_encode_block, Scalar, DIMS)(stream, bundle->dataArr);
   zfp_stream_flush(stream);
 
-  UInt checksum = hashBitstream(stream_data(s), stream_size(s));
+  uint64 checksum = hashBitstream(stream_data(s), stream_size(s));
   assert_int_equal(checksum, CHECKSUM_ENCODED_BLOCK);
 }
